Loading feature complete

+++ pages/api/[id].js
@@ -1,4 +1,4 @@
-import { UpdateOutlined } from "@mui/icons-material";
+import { ConstructionOutlined, UpdateOutlined } from "@mui/icons-material";
 import { ObjectID } from "bson";
 import { withTheme } from "styled-components";
 
@@ -15,9 +15,7 @@ export default async function (req, res){
   const positions = []
   const range = (start, end, length = end - start + 1) =>
   Array.from({ length }, (_, i) => start + i)
-
-
-  let { position } = body
+  let { position, name, phone, action } = body
 
   clientList.map(client => {
     positions.push(client.position)
@@ -27,27 +25,43 @@ export default async function (req, res){
   }, -Infinity)
 
   if(req.method === "POST") {
-    if(clientList.length == [].length) {
-      body.position = 0
-    }
-    else {
-      const positions = []
-      clientList.map(client => {
-        positions.push(client.position)
-      })
-      
-      body.position = highestPosition + 1
+      if(clientList.length == [].length) {
+        body.position = 0
       }
-    await clients.insertOne(body)
+      else {
+        const positions = []
+        clientList.map(client => {
+          positions.push(client.position)
+        })
+        
+        body.position = highestPosition + 1
+        }
+      await clients.insertOne(body)
     }
   
+ if(req.method === "PUT") {
+    if (body.action === "goDown") {
+      const currentClient = await clients.findOne({position: position})
+      const beforeClient = await clients.findOne({position: position+1})
+      await clients.updateOne({_id: currentClient._id}, {$inc: {position: 1}})
+      await clients.updateOne({_id: beforeClient._id}, {$inc:{position: -1}})
+    }
+    else if (body.action === "goUp") {
+      const currentClient = await clients.findOne({position: position})
+      const nextClient = await clients.findOne({position: position-1})
+      await clients.updateOne({_id: currentClient._id}, {$inc: {position: -1}})
+      await clients.updateOne({_id: nextClient._id}, {$inc:{position: 1}})
+      
+    }
+  }
+
   if(req.method === "DELETE") {
     await clients.deleteOne({position: position})
 
     await clients.updateMany({position: {$gt: position}}, {$inc:{position: -1}})
   }
 
-  const response = await clients.find({}).toArray()
+  const response = await clients.find({}).sort({position: 1}).toArray()
   await client.close()
   res.status(200).json(response)
 }
